requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
PyYAML>=6.0.0

# Google AI
google-generativeai>=0.3.0

# Development
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.1
mypy>=0.930

# Documentation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0

# CLI Enhancements
click>=8.0.0
rich>=10.0.0
